refactor(st-09053): split connection runtime modules
