+++ .pr-body-st-09053.md
@@ -1,17 +1,23 @@
 ## Summary
-- modularize the oversized relational connection-manager facade into a stable public facade plus focused runtime helpers
+- modularize the oversized relational connection-manager facade into a stable public facade plus focused runtime modules
 - split the monolithic connection-manager and lifecycle test cluster into focused connection-manager test modules
 - preserve connection-manager behavior, lifecycle semantics, vendor handling, and public imports while improving reviewability
 
 ## What Changed
-- extracted runtime helpers into `packages/tools/src/data/relational/connection/connection-manager-runtime.ts`
 - reduced `packages/tools/src/data/relational/connection/connection-manager.ts` from `640` lines to `255` lines
+- split the former runtime helper blob into focused modules:
+  - `connection-initialization.ts`
+  - `connection-cleanup.ts`
+  - `connection-health.ts`
+  - `connection-manager-runtime-types.ts`
+  - `connection-manager-runtime.ts` as a small export barrel
 - replaced the old monolithic test cluster with focused suites under `packages/tools/tests/data/relational/connection/`
 - added shared test harness helpers for mocked driver and SQLite runtime coverage
 - recorded the modularization evidence in `docs/st09053-connection-manager-modularization.md`
 
 ## Acceptance Criteria Coverage
 - `connection-manager.ts` is now below the `300` line modularization cutoff
+- the extracted runtime logic is also split into focused modules instead of a single oversized helper
 - initialization, cleanup, health, SQLite non-query detection, and pool-metric runtime logic were extracted behind the same public facade
 - the oversized connection-manager test cluster was split into focused lifecycle, operations, config, runtime, and edge-case suites
 - public `ConnectionManager` imports and lifecycle/query behavior were preserved
